Move default result storage CLI to top level
