+++ tests/cli/test_result_storage.py
@@ -7,28 +7,28 @@ from prefect.testing.cli import invoke_and_assert
 from prefect.utilities.asyncutils import run_sync_in_worker_thread
 
 
-async def test_server_storage_cli_e2e(
+async def test_result_storage_cli_e2e(
     enable_ephemeral_server,
     prefect_client: PrefectClient,
     tmp_path,
 ):
     await run_sync_in_worker_thread(
         invoke_and_assert,
-        ["server", "storage", "inspect"],
+        ["result-storage", "inspect"],
         expected_code=0,
-        expected_output_contains="No server default result storage is configured.",
+        expected_output_contains="No default result storage is configured.",
     )
 
     await run_sync_in_worker_thread(
         invoke_and_assert,
-        ["server", "storage", "inspect", "--output", "yaml"],
+        ["result-storage", "inspect", "--output", "yaml"],
         expected_code=1,
         expected_output_contains="Only 'json' output format is supported.",
     )
 
     await run_sync_in_worker_thread(
         invoke_and_assert,
-        ["server", "storage", "set", "server-default-results"],
+        ["result-storage", "set", "server-default-results"],
         expected_code=1,
         expected_output_contains=(
             "'server-default-results' is not valid. Slug must contain a '/'"
@@ -44,9 +44,9 @@ async def test_server_storage_cli_e2e(
     try:
         await run_sync_in_worker_thread(
             invoke_and_assert,
-            ["server", "storage", "set", f"local-file-system/{block_name}"],
+            ["result-storage", "set", f"local-file-system/{block_name}"],
             expected_code=0,
-            expected_output_contains="Set server default result storage.",
+            expected_output_contains="Set default result storage.",
         )
 
         configuration = await prefect_client.read_server_default_result_storage()
@@ -54,7 +54,7 @@ async def test_server_storage_cli_e2e(
 
         result = await run_sync_in_worker_thread(
             invoke_and_assert,
-            ["server", "storage", "inspect", "--output", "json"],
+            ["result-storage", "inspect", "--output", "json"],
             expected_code=0,
         )
         assert json.loads(result.stdout.strip()) == {
@@ -64,9 +64,9 @@ async def test_server_storage_cli_e2e(
 
         await run_sync_in_worker_thread(
             invoke_and_assert,
-            ["server", "storage", "clear"],
+            ["result-storage", "clear"],
             expected_code=0,
-            expected_output_contains="Cleared server default result storage.",
+            expected_output_contains="Cleared default result storage.",
         )
 
         configuration = await prefect_client.read_server_default_result_storage()
@@ -74,9 +74,9 @@ async def test_server_storage_cli_e2e(
 
         await run_sync_in_worker_thread(
             invoke_and_assert,
-            ["server", "storage", "set", "--id", str(block_document_id)],
+            ["result-storage", "set", "--id", str(block_document_id)],
             expected_code=0,
-            expected_output_contains="Set server default result storage.",
+            expected_output_contains="Set default result storage.",
         )
 
         configuration = await prefect_client.read_server_default_result_storage()
